// backend/src/types/generated/public/Messages.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ConversationsId } from './Conversations';

/** Identifier type for public.messages */
export type MessagesId = string & { __brand: 'public.messages' };

/** Represents the table public.messages */
export default interface Messages {
  id: MessagesId;

  conversation_id: ConversationsId;

  sender_type: string;

  sender_id: string | null;

  sender_name: string | null;

  message_type: string | null;

  content: string;

  external_message_id: string | null;

  client_message_id: string | null;

  reply_to_message_id: MessagesId | null;

  is_forwarded: boolean | null;

  forwarded_from_name: string | null;

  delivery_status: string | null;

  delivered_at: string | null;

  read_at: string | null;

  is_read: boolean | null;

  metadata: Record<string, unknown> | null;

  created_at: string | null;

  legacy_message_id: string | null;

  event_type: string | null;

  attachment_url: string | null;
}

/** Represents the initializer for the table public.messages */
export interface MessagesInitializer {
  /** Default value: gen_random_uuid() */
  id?: MessagesId;

  conversation_id: ConversationsId;

  sender_type: string;

  sender_id?: string | null;

  sender_name?: string | null;

  /** Default value: 'text'::character varying */
  message_type?: string | null;

  content: string;

  external_message_id?: string | null;

  client_message_id?: string | null;

  reply_to_message_id?: MessagesId | null;

  /** Default value: false */
  is_forwarded?: boolean | null;

  forwarded_from_name?: string | null;

  /** Default value: 'accepted'::character varying */
  delivery_status?: string | null;

  delivered_at?: string | null;

  read_at?: string | null;

  /** Default value: false */
  is_read?: boolean | null;

  metadata?: Record<string, unknown> | null;

  /** Default value: now() */
  created_at?: string | null;

  legacy_message_id?: string | null;

  event_type?: string | null;

  attachment_url?: string | null;
}

/** Represents the mutator for the table public.messages */
export interface MessagesMutator {
  id?: MessagesId;

  conversation_id?: ConversationsId;

  sender_type?: string;

  sender_id?: string | null;

  sender_name?: string | null;

  message_type?: string | null;

  content?: string;

  external_message_id?: string | null;

  client_message_id?: string | null;

  reply_to_message_id?: MessagesId | null;

  is_forwarded?: boolean | null;

  forwarded_from_name?: string | null;

  delivery_status?: string | null;

  delivered_at?: string | null;

  read_at?: string | null;

  is_read?: boolean | null;

  metadata?: Record<string, unknown> | null;

  created_at?: string | null;

  legacy_message_id?: string | null;

  event_type?: string | null;

  attachment_url?: string | null;
}